opencv/requirements.txt
diffusers
transformers
xformers
accelerate
opencv-contrib-python
controlnet_aux
google-generativeai
gradio
Pillow
torch
